migrate serve_flask_app to only remaining using module
